n7-strikers/n7_strikers/agent_runtime/service.py

import asyncio
import logging
import aiohttp
from .config import settings

logger = logging.getLogger("n7-striker.agent-runtime")

class AgentRuntimeService:
    """
    Agent Runtime Service.
    Responsibility: Handle registration, heartbeat, and auth with Core.
    """
    def __init__(self):
        self._running = False
        self._session = None
        self._token = None

    async def start(self):
        self._running = True
        logger.info("AgentRuntimeService started.")
        self._session = aiohttp.ClientSession()
        await self._authenticate()
        asyncio.create_task(self._heartbeat_loop())

    async def stop(self):
        self._running = False
        if self._session:
            await self._session.close()
        logger.info("AgentRuntimeService stopped.")

    async def _authenticate(self):
        """
        Authenticates with Core.
        """
        try:
            logger.info("Authenticating with Core...")
            self._token = "mock-jwt-token"
        except Exception as e:
            logger.error(f"Authentication failed: {e}")

    async def _heartbeat_loop(self):
        """
        Sends periodic heartbeats.
        """
        while self._running:
            # TODO: Send heartbeat
            await asyncio.sleep(30)
