n7-strikers/requirements.txt

pydantic>=2.0.0
nats-py>=2.6.0
typer>=0.9.0
psutil
cryptography
pydantic-settings>=2.0.0\naiohttp>=3.9.0
